% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_s0.R
\name{choose_s0}
\alias{choose_s0}
\title{Select s0 using the method of Tusher, Tibshirani and Chu (2001)}
\usage{
choose_s0(beta, sds)
}
\arguments{
\item{beta}{Effect size estimates}

\item{sds}{Estimates of sqrt(var(beta))}
}
\description{
Select s0 using the method of Tusher, Tibshirani and Chu (2001)
}
