% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/choose_z_even.R
\name{choose_z_even}
\alias{choose_z_even}
\title{Select threshold value in each segment for a range of lambda values}
\usage{
choose_z_even(perm.stats, nlam, bw, pos, z0, x.range = NULL,
  max.log.lam = NULL, seg.ends = NULL, return.z = TRUE,
  keep.lists = FALSE, smooth.func = c("ksmooth", "ksmooth_0"))
}
\arguments{
\item{perm.stats}{p x B matrix of permutation statistics}

\item{nlam}{Number of lambda values to consider}

\item{bw}{Binwidth for smoothing}

\item{pos}{Positions}

\item{z0}{Reference level for merging}

\item{max.log.lam}{Maximum value of log(lambda) to consider}

\item{seg.ends}{List of segment ends. If NULL assume there is only one segment}

\item{return.z}{if FALSE only return maxes list}

\item{keep.lists}{If return.z=TRUE also return maxes list}

\item{smooth.func}{Either ksmooth or ksmooth_0}
}
\value{
If return.z =TRUE, return a matrix that is nlam x (nseg + 1). The first column is log.lambda.
Subsequent columns give the threshold for each segment. If keep.lists=TRUE or return.z=FALSE (also) return
a list of cluster peaks for each segment. mx has length nseg and each element has two columns - peak height and
empirical lambda value.
}
\description{
Selects a series of threshold values so that each segment has approximately the same false cluster rate
}

