% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_fret_stats.R
\name{collect_fret_stats}
\alias{collect_fret_stats}
\title{Collect fret_stats temporary files}
\usage{
collect_fret_stats(temp.dir, temp.prefix, which.chunk, n.perm = NULL,
  out.file = NULL, del.temp = FALSE, min.interval.width = NULL)
}
\arguments{
\item{temp.dir}{Directory containing temporary files}

\item{temp.prefix}{Prefix of temporary files (e.g. If your files are named
test_chr1.N.RData where N is the chunk number then temp.prefix is test_chr1)}

\item{which.chunk}{Vector of integers. Which chunks to collect}

\item{n.perm}{Temporary parameter for compatibility with previous versions.}

\item{out.file}{Name an output file}

\item{del.temp}{Delete the temporary files when done?}

\item{min.interval.width}{Minimum interval length when automatically determining interval endpoints}
}
\description{
Combine temporary files from fret_stats into one file.
Also determine interval boundaries.
}
