% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveQTL_preprocess_funcs.R
\name{corrected_forCovariates}
\alias{corrected_forCovariates}
\title{Correct for covariates.}
\usage{
corrected_forCovariates(x, Covariates)
}
\arguments{
\item{x}{a vector of length N (# of samples) containing data.}

\item{Covariates}{a matrix (or a vector if M = 1) with N by M (# of covariates)
containing covariates to correct for.}
}
\value{
a vector of length N; covariates corrected data.
}
\description{
This function corrects for covariates.
}

