% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase1_test_windows.R
\name{dnase1_test_windows}
\alias{dnase1_test_windows}
\title{Run a variety of tests in pre-defined windows}
\usage{
dnase1_test_windows(dat.file, pheno.file, window.file, chr, maxit = 50,
  trait = "pheno", names = "name", save.collapsed = NULL,
  waveqtl.min.pval = 1e-08, waveQTL_loc = "~/WaveQTL-master/bin/WaveQTL")
}
\arguments{
\item{dat.file}{Data file}

\item{pheno.file}{Phenotype file}

\item{maxit}{Maximum iterations for Huber estimation}

\item{trait}{Trait name in pheno.file}

\item{names}{Names column in pheno.file}
}
\value{
A list
}
\description{
Calculate test statistics binning over regions
}
