% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_segments.R
\name{find_segments}
\alias{find_segments}
\title{Use variance of smoothed  permutation test statistics to define interval endpoints}
\usage{
find_segments(vv, pos, min.length, z0 = NULL, z = NULL)
}
\arguments{
\item{vv}{Variance of smoothed permutation test statistics}

\item{pos}{Position vector. Should be the same length as vv}

\item{min.length}{Minimum interval length}

\item{z0,}{z Two thresholds used to define ``high variance regions''.}
}
\value{
2 by K matrix listing interval endpoints.
}
\description{
Use variance of smoothed  permutation test statistics to define interval endpoints.
Intervals are placed so that high variance regions are separated from each other and from low variance
regions.
}
\details{
Currently, find_segments will smooth vv using a bandwidth of min.length/2.
If you don't supply z and z0, z will be the 75th percentlie of these smoothed values and z0 will be
the 25th percentile.
}

