% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveQTL_preprocess_funcs.R
\name{fiter.WCs}
\alias{fiter.WCs}
\title{Filter low count WCs.}
\usage{
fiter.WCs(Data, meanR.thresh)
}
\arguments{
\item{Data}{matrix (or a vector when N = 1) with N (# of samples) by T (# of bps in a region); This matrix contains original data before a wavelet transform; Here, T should be a power of 2.}

\item{meanR.thresh}{If average reads across individuals <= meanR.thresh,
those WCs are filtered out.}
}
\value{
filtered.WCs a vector of length T; t-th element indicates whether t-th WC in output from the function \code{\link{FWT}} is filtered (0) or not (1).
}
\description{
This function filters out some WCs that are computed based on very
low counts. This function considers WCs as low count if the total
counts used in their computation were less than or equal to `meanR.thresh'
per individual on average.
}

