% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fret_choose_z2.R
\name{fret_choose_z2}
\alias{fret_choose_z2}
\title{Find thresholds for a range of lambda values. Calculate FDR.}
\usage{
fret_choose_z2(max1.list, perm.maxes.list, nbp, zmin, fdr.max = 0.8,
  seg.names = NULL, target.fdr = NULL)
}
\arguments{
\item{max1.list}{List of peak heights for each segment.}

\item{perm.maxes.list}{a list of matrices giving peak heights for permutations. Each matrix has two columns. First column is the threshold level,
second column is lambda per bp (num of peaks less than threshold divided by number of permutations and number of base-pairs).}

\item{nbp}{Number of base-pairs per segment}

\item{zmin}{}

\item{fdr.max}{}

\item{seg.names}{}

\item{target.fdr}{}
}
\value{
A list with items z, Robs, and fdr.
}
\description{
Find thresholds for a range of lambda values. Calculate FDR.
}

