% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_rates_prelim.R
\name{fret_rates_prelim}
\alias{fret_rates_prelim}
\title{Calculate per-base error rates for original and permutation peaks}
\usage{
fret_rates_prelim(fret.obj, segment.bounds = NULL, parallel = FALSE,
  save.file = NULL)
}
\arguments{
\item{fret.obj}{Either a fret object or a file containing a fret object}

\item{segment.bounds}{Optional data frame of segment bounds. If ommitted
then fret.obj should contain an item `seg.bounds` which will be used instead.}

\item{parallel}{Run in parallel?}
}
\description{
Calculate per-base error rates for original and permutation peaks
}

