% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_stats4.R
\name{fret_stats}
\alias{fret_stats}
\title{Calculate FRET test statistics}
\usage{
fret_stats(pheno.file, trait.file, s0, seed, n.perm, zmin = NULL, z0 = zmin
  * 0.3, pheno.transformation = NULL, trait = c("x"), covariates = c(),
  bandwidth = 150, smoother = c("ksmooth_0", "ksmooth", "none"),
  stat.type = c("huber", "lm"), maxit = 50, margin = 5 * bandwidth,
  range = NULL, chunksize = 1e+05, which.chunks = NULL,
  temp.prefix = NULL, temp.dir = "./", chrom = "chr1", parallel = TRUE,
  cores = parallel::detectCores() - 1, digits = Inf)
}
\arguments{
\item{pheno.file}{File containing genomic phenotype}

\item{trait.file}{File containing trait information}

\item{s0}{Variance inflation constant}

\item{seed}{Seed}

\item{n.perm}{Number of permutations.
If n.perm=0, only test statistics for unpermuted data will be calculated}

\item{zmin}{Minimum threshold (may be missing if n.perm=0)}

\item{z0}{Merging threshold. If missing z0 = 0.3*zmin.}

\item{pheno.transformation}{If the phenotype is to be transformed, provide a function taking one argument}

\item{trait}{Name of trait (should match header of trait.file)}

\item{covariates}{List of covariates found in trait.file to adjust for}

\item{bandwidth}{Smoothing bandwidth}

\item{stat.type}{Either "huber" or "lm"}

\item{maxit}{Maximum iterations for Huber estimator.}

\item{range}{Base-pair range. If missing, stats for the whole file will be calculated}

\item{chunksize}{Size of chunks to read at one time}

\item{chrom}{Name of chromosome}

\item{save.perm.stats}{Should the permutation test statistics be saved?}

\item{smooother}{Either "ksmooth_0" or "ksmooth". The ksmooth_0 smoother treats base-pairs
with no data as though the phenotype at that position is equal to zero for all samples (appropriate for DNase-seq data).
The ksmooth smoother treats these base-pairs as missing (appropriate for bisulfite sequencing data).}

\item{out.file}{Name an output file}
}
