% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fret_choose_z3.R
\name{fret_step2}
\alias{fret_step2}
\title{Calculate lambda and fdr for each peak}
\usage{
fret_step2(max1, max.perm, n.perm, zmin, segment.bounds, fdr.max = 0.8)
}
\arguments{
\item{max1}{Table giving peak height, position, and chromosome for original data}

\item{max.perm}{Table giving peak height, position, and chromosome for permuted data}

\item{n.perm}{Number of permutations}

\item{zmin}{Minimum Threshold (length 1 or 2)}

\item{segment.bounds}{Data frame with three columns. First column is chromosome.
Columns two and three are segment start and stop}

\item{fdr.max}{Maximum fdr to keep data for}

\item{target.fdr}{Target fdr values (vector)}
}
\value{
A list with items z, Robs, and fdr.
}
\description{
Find thresholds for a range of lambda values. Calculate FDR.
}

