% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_thresholds.R
\name{fret_thresholds}
\alias{fret_thresholds}
\title{Get thresholds for a target FDR level}
\usage{
fret_thresholds(obj, target.fdr, stats.files)
}
\arguments{
\item{obj}{Object produced by fret_rates}

\item{target.fdr}{trget fdr level}

\item{stats.files}{Two column matrix.
The first column is chromosome. The second column is the corresponding stats file.}
}
\value{
An object that can be passed to fret_thresholds
}

