% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveQTL_preprocess_funcs.R
\name{FWT}
\alias{FWT}
\title{Perform a wavelet transform.}
\usage{
FWT(Data, filter.number = 1, family = "DaubExPhase")
}
\arguments{
\item{Data}{matrix (or a vector when N = 1) with N (# of samples) by T (# of bps in a region);
This matrix contains original data to be decomposed; Here, T should be a power of 2.}

\item{filter.number}{default=1; argument to the function \code{\link{wd}} in the R package
\code{\link{wavethresh}}; See their manual for details.}

\item{family}{default="DaubExPhase"; argument to the function \code{\link{wd}} in
the R package \code{\link{wavethresh}}; See their manual for details.}
}
\value{
WCs a matrix with N (# of samples) by T (# of bps in a region); n-th row contains WCs
for n-th sample; WCs are ordered from low resolution WC to high resolution WC; For example,
with a Haar wavelet transform, the first column contains WC (precisely speaking, scaling
coefficient) that corresponds to sum of data in the region. The second column contains WC
that contrasts the data in the first half vs second half of the region. The last column
contains WC that contrasts the data in the (T-1)-th bp vs T-th bp.
}
\description{
This function performs a wavelet transform using a \code{\link{wavethresh}} R package
and returns WCs in the order that corresponds to output from the function
\code{\link{fiter.WCs}}. For now, the function doesn't allow users to specify the level of wavelet
decomposition and uses the maximum level decomposition.
}

