% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Helper function for calculating test statistics}
\usage{
get_stats(file_name, nchunks, chunksize, margin, X, trait, covariates, sample,
  pheno_transformation = NULL, cores, libs, s0, stat_fun, resid_fun,
  chunks = "all")
}
\arguments{
\item{file_name}{Name of phenotyp file}

\item{nchunks}{Number of chunks in file}

\item{chunksize}{Chunk size}

\item{margin}{Margin}

\item{X}{Data frame of trait information}

\item{trait}{String; name of trait in X}

\item{covariates}{Vector of names of covariates in X}

\item{sample}{String; name of sample name in X}

\item{pheno_transformation}{Phenotype transformation}

\item{cores}{Number of cores}

\item{s0}{s0}

\item{stat_fun}{Function to calculate test stats}

\item{resid_fun}{Function to calculate residual}

\item{chunks}{Which chunks; may only be "all" or contiguous integers}
}
