% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{huber_stats}
\alias{huber_stats}
\title{Calculate Huber test statistics}
\usage{
huber_stats(Y, x, s0 = 0, k = 1.345, maxit = 50)
}
\arguments{
\item{Y}{matrix (p x n)}

\item{x}{trait values}

\item{s0}{Additional variance}

\item{k}{Threshold for huber estimator in multiples of scale parameter.}

\item{maxit}{Maixum iterations to pass to rlm.}
}
\value{
3 by p matrix giving coefficient estimates, sd estimates and statistic (including s0 adjustment)
}
\description{
Calculate two sample Huber
}

