% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ksmooth_0_cpp}
\alias{ksmooth_0_cpp}
\title{Kernel smoother for integer locations}
\usage{
ksmooth_0_cpp(x, y, xout, bandwidth)
}
\arguments{
\item{x}{Input locations (integers), sorted}

\item{y}{Input observations, should be length of x and non-missing}

\item{xout}{Output locations}

\item{bandwidth}{Bandwidth. Should be an odd integer}
}
\description{
Assumes missing locations have observation of 0
}
