% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksmooth_0.R
\name{ksmooth_0_jean}
\alias{ksmooth_0_jean}
\title{Box kernel smoother for integer positions}
\usage{
ksmooth_0_jean(x, y, xout, bandwidth, chunksize)
}
\arguments{
\item{x}{Vector of positions (integers)
#'@param y Vector of observations to be smoothed}

\item{xout}{Vector of output positions}

\item{bandwidth}{Smoother bandwidth. Should be an odd integer}

\item{chunksize}{Size of chunks to break xout up into. Speeds computation.}
}
\value{
Vector of smoothed values with length equal to length of xout.
}
\description{
Box kernel for use with DNase-seq and similar data. Assumes missing integer positions have observations of 0.
}
