% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ksmooth_0.R
\name{ksmooth_0_old}
\alias{ksmooth_0_old}
\title{Kernel smoother assuming missing positions are 0}
\usage{
ksmooth_0_old(x, y, xout, bandwidth)
}
\arguments{
\item{x}{Vector of positions}

\item{y}{Vector of test statistics}

\item{bandwidth}{Smoother bandwidth}

\item{maxit}{Maixum iterations to pass to rlm.}
}
\value{
2 by p matrix. Top row is coefficient estimate. Bottom row is sd estimates.
}
\description{
Kernel smoother assuming missing positions are 0
}

