% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{lm_stats_parallel}
\alias{lm_stats_parallel}
\title{Calculate OLS test statistics using parallel package}
\usage{
lm_stats_parallel(Y, x, cores, s0)
}
\arguments{
\item{Y}{matrix (p x n)}

\item{x}{trait values}

\item{cores}{Number of cores to use}

\item{s0}{Additional variance}
}
\value{
3 by p matrix giving coefficient estimates, sd estimates and statistic (including s0 adjustment)
}
\description{
Calculate OLS test statistics using parallel package
}
