% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxes_1seg.R
\name{maxes_1seg}
\alias{maxes_1seg}
\title{Get peak heights and permuted peak heights for one segment}
\usage{
maxes_1seg(dat.file, pheno.file, s0, zmin, seed, n.perm,
  save.perm.stats = FALSE, segment.bounds = NULL, out.file = NULL,
  max1.file = NULL, z0 = zmin * 0.3, bandwidth = 50, maxit = 50)
}
\arguments{
\item{dat.file}{Data file. Should have first
column as position and a header giving sample names. One row per base pair.}

\item{pheno.file}{Phenotype file. First column should correspond to the header of dat.file.}

\item{s0}{Variance inflation constant.}

\item{zmin}{Lower bound on significance thresholds.}

\item{seed}{Set a seed for running permutations (Required)}

\item{n.perm}{Number of permutations}

\item{save.perm.stats}{Do you want to save all the permutation statistics?}

\item{segment.bounds}{Length 2 vector giving start and stop base-pair for the segment.
If NULL, all base-pairs in dat.file are used.}

\item{out.file}{Output file (optional). If NULL will append "_mx.RData" to the leading portion of the
dat.file name.}

\item{max1.file}{File with data from running with n.perm=0}

\item{z0}{Reference level for merging}

\item{bandwidth}{Smoother bandwidth.}

\item{maxit}{Maximum iterations for rlm.}
}
\value{
Saves an RData file with maxes list and other info. Returns the length of that list.
}
\description{
Finds peak heights and permuatation peak heights for one segment
}

