% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveQTL_preprocess_funcs.R
\name{Normalize.WCs}
\alias{Normalize.WCs}
\title{Normalize WCs.}
\usage{
Normalize.WCs(WCs, Covariates = NULL)
}
\arguments{
\item{WCs}{a matrix with N (# of samples) by T (# of bps in a region or # of WCs);
n-th row contains WCs for n-th sample.}

\item{Covariates}{default = NULL; a matrix (or a vector if M = 1) with N by M
(# of covariates) containing covariates to correct for.}
}
\value{
QT_WCs a matrix with N (# of samples) by T (# of bps in a region or # of WCs);
It contains normalized WCs (Quantile-transformed and covariate-corrected WCs).
}
\description{
This function quantile-transforms WCs to a standard normal distribution.
If covariates are provided, it corrects the quantile-transformed WCs for the covariates
and quantile-transforms the covariates-corrected WCs to a standard normal distribution.
}

