% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveQTL_preprocess_funcs.R
\name{QT_randomTie}
\alias{QT_randomTie}
\title{Quantile-transform data to a standard normal distribution.}
\usage{
QT_randomTie(x)
}
\arguments{
\item{x}{a vector containing data to be quantile-transformed.}
}
\value{
quantile-transformed data.
}
\description{
This function quantile-transforms data to a standard normal distribution. It randomly assign
ranks for ties.
}

