% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_waveQTL.R
\name{run_waveQTL}
\alias{run_waveQTL}
\title{Run waveQTL}
\usage{
run_waveQTL(dat, x, waveQTL_loc = "~/WaveQTL-master/bin/WaveQTL")
}
\arguments{
\item{dat}{matrix or data frame of data. First column should be position.
second column should be window number, the rest should be data.}

\item{x}{Phenotype}

\item{waveQTL_loc}{Path to the waveQTL program}

\item{names}{Names column in pheno.file}
}
\value{
A list
}
\description{
Runs the method of Shim and Stephens 2014
}

