% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_fret_stats.R
\name{stats_to_rates}
\alias{stats_to_rates}
\title{Collect fret_stats temporary files}
\usage{
stats_to_rates(file_list, seg_type = c("by_file", "by_chromosome", "find",
  "provided"), chromosome, segment_bounds, min_segment_length = NULL)
}
\arguments{
\item{file_list}{List of files produced by fret_stats}

\item{seg_type}{One of "by_file", "by_chromosome", "find", "provided". See description.}

\item{chromosome}{Vector the same length as file_list giving the chromosome for each file.
This is only required if seg_type is not equal to by_file.}

\item{segment_bounds}{Optional data frame with columns start, stop and chrom.}

\item{min_segment_length}{Used if seg_type="find". Minimum segment length.
Defaults to 50*bandwidth if not provided.}
}
\description{
Combine temporary files from fret_stats into one file.
Also determine interval boundaries.
}
