% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_segments.R
\name{find_segments}
\alias{find_segments}
\title{Use variance of smoothed  permutation test statistics to define interval endpoints}
\usage{
find_segments(vv, pos, min_length, q = 0.05)
}
\arguments{
\item{vv}{Variance of smoothed permutation test statistics}

\item{pos}{Position vector. Should be the same length as vv}

\item{min_length}{Minimum interval length}

\item{q}{Quantile for upper limit for definition of high variance region.}
}
\value{
2 by K matrix listing interval endpoints.
}
\description{
Use variance of smoothed  permutation test statistics to define interval endpoints.
Intervals are placed so that high variance regions are separated from each other and from low variance
regions.
}
