% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fret.R
\name{fret_choose_z}
\alias{fret_choose_z}
\title{Find thresholds for a range of lambda values. Calculate FDR.}
\usage{
fret_choose_z(file.list, zmin, nlam, log.lambda.min = NULL,
  log.lambda.max = NULL)
}
\arguments{
\item{file.list}{List of files. Each file should contain a list of objects produced by
discopony_maxes1}

\item{zmin}{Lower bound on significance thresholds.}

\item{nlam}{Number of lambda values to consider}
}
\value{
A list with items z, Robs, and fdr.
}
\description{
Find thresholds for a range of lambda values. Calculate FDR.
}

