% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_rates.R
\name{fret_rates}
\alias{fret_rates}
\title{Calculate lambda and fdr for each peak}
\usage{
fret_rates(file.list, fdr.max = 0.8)
}
\arguments{
\item{file.list}{List of files with output from fret_rates_prelim}

\item{fdr.max}{Maximum fdr to keep data for}
}
\value{
An object that can be passed to fret_thresholds
}
\description{
Find thresholds for a range of lambda values. Calculate FDR.
}

