% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_rates_prelim.R
\name{fret_rates_prelim}
\alias{fret_rates_prelim}
\title{Calculate per-base error rates for original and permutation peaks}
\usage{
fret_rates_prelim(fret_obj, segment_lengths, segment_bounds)
}
\arguments{
\item{fret_obj}{fret object or a file name or a vector of file names.}

\item{segment_lengths}{data frame with (at least) two columns: segment and length in base-pairs}

\item{segment_bounds}{data frame with (at least) two columns: start and stop base pair position}
}
\description{
Calculate per-base error rates for original and permutation peaks
}
